refactor: find preve heading pos
